Move some more classes to namespace
