+++ library/SimplePie/Copyright.php
@@ -41,86 +41,15 @@
  * @license http://www.opensource.org/licenses/bsd-license.php BSD License
  */
 
-/**
- * Manages `<media:copyright>` copyright tags as defined in Media RSS
- *
- * Used by {@see SimplePie_Enclosure::get_copyright()}
- *
- * This class can be overloaded with {@see SimplePie::set_copyright_class()}
- *
- * @package SimplePie
- * @subpackage API
- */
-class SimplePie_Copyright
-{
-	/**
-	 * Copyright URL
-	 *
-	 * @var string
-	 * @see get_url()
-	 */
-	var $url;
+use SimplePie\Copyright;
 
-	/**
-	 * Attribution
-	 *
-	 * @var string
-	 * @see get_attribution()
-	 */
-	var $label;
+class_exists('SimplePie\Copyright');
 
-	/**
-	 * Constructor, used to input the data
-	 *
-	 * For documentation on all the parameters, see the corresponding
-	 * properties and their accessors
-	 */
-	public function __construct($url = null, $label = null)
+// @trigger_error(sprintf('Using the "SimplePie_Copyright" class is deprecated since SimplePie version 1.x, use "SimplePie\Copyright" instead.'), \E_USER_DEPRECATED);
+
+if (\false) {
+	/** @deprecated since SimplePie 1.x, use "SimplePie\Copyright" instead */
+	class SimplePie_Copyright extends Copyright
 	{
-		$this->url = $url;
-		$this->label = $label;
-	}
-
-	/**
-	 * String-ified version
-	 *
-	 * @return string
-	 */
-	public function __toString()
-	{
-		// There is no $this->data here
-		return md5(serialize($this));
-	}
-
-	/**
-	 * Get the copyright URL
-	 *
-	 * @return string|null URL to copyright information
-	 */
-	public function get_url()
-	{
-		if ($this->url !== null)
-		{
-			return $this->url;
-		}
-
-		return null;
-	}
-
-	/**
-	 * Get the attribution text
-	 *
-	 * @return string|null
-	 */
-	public function get_attribution()
-	{
-		if ($this->label !== null)
-		{
-			return $this->label;
-		}
-
-		return null;
 	}
 }
-
-class_alias('SimplePie_Copyright', 'SimplePie\Copyright', false);
