+++ library/SimplePie/Exception.php
@@ -41,13 +41,15 @@
  * @license http://www.opensource.org/licenses/bsd-license.php BSD License
  */
 
-/**
- * General SimplePie exception class
- *
- * @package SimplePie
- */
-class SimplePie_Exception extends Exception
-{
-}
+use SimplePie\Exception as SimplePieException;
 
-class_alias('SimplePie_Exception', 'SimplePie\Exception', false);
+class_exists('SimplePie\Exception');
+
+// @trigger_error(sprintf('Using the "SimplePie_Exception" class is deprecated since SimplePie version 1.x, use "SimplePie\Exception" instead.'), \E_USER_DEPRECATED);
+
+if (\false) {
+	/** @deprecated since SimplePie 1.x, use "SimplePie\Exception" instead */
+	class SimplePie_Exception extends SimplePieException
+	{
+	}
+}
