+++ src/Category.php
@@ -46,10 +46,124 @@
 
 namespace SimplePie;
 
-class_exists('SimplePie_Category');
+/**
+ * Manages all category-related data
+ *
+ * Used by {@see \SimplePie\Item::get_category()} and {@see \SimplePie\Item::get_categories()}
+ *
+ * This class can be overloaded with {@see \SimplePie\SimplePie::set_category_class()}
+ *
+ * @package SimplePie
+ * @subpackage API
+ */
+class Category
+{
+	/**
+	 * Category identifier
+	 *
+	 * @var string|null
+	 * @see get_term
+	 */
+	var $term;
 
-if (\false) {
-	class Category extends \SimplePie_Category
+	/**
+	 * Categorization scheme identifier
+	 *
+	 * @var string|null
+	 * @see get_scheme()
+	 */
+	var $scheme;
+
+	/**
+	 * Human readable label
+	 *
+	 * @var string|null
+	 * @see get_label()
+	 */
+	var $label;
+
+	/**
+	 * Category type
+	 *
+	 * category for <category>
+	 * subject for <dc:subject>
+	 *
+	 * @var string|null
+	 * @see get_type()
+	 */
+	var $type;
+
+	/**
+	 * Constructor, used to input the data
+	 *
+	 * @param string|null $term
+	 * @param string|null $scheme
+	 * @param string|null $label
+	 * @param string|null $type
+	 */
+	public function __construct($term = null, $scheme = null, $label = null, $type = null)
 	{
+		$this->term = $term;
+		$this->scheme = $scheme;
+		$this->label = $label;
+		$this->type = $type;
+	}
+
+	/**
+	 * String-ified version
+	 *
+	 * @return string
+	 */
+	public function __toString()
+	{
+		// There is no $this->data here
+		return md5(serialize($this));
+	}
+
+	/**
+	 * Get the category identifier
+	 *
+	 * @return string|null
+	 */
+	public function get_term()
+	{
+		return $this->term;
+	}
+
+	/**
+	 * Get the categorization scheme identifier
+	 *
+	 * @return string|null
+	 */
+	public function get_scheme()
+	{
+		return $this->scheme;
+	}
+
+	/**
+	 * Get the human readable label
+	 *
+	 * @param bool $strict
+	 * @return string|null
+	 */
+	public function get_label($strict = false)
+	{
+		if ($this->label === null && $strict !== true)
+		{
+			return $this->get_term();
+		}
+		return $this->label;
+	}
+
+	/**
+	 * Get the category type
+	 *
+	 * @return string|null
+	 */
+	public function get_type()
+	{
+		return $this->type;
 	}
 }
+
+class_alias('SimplePie\Category', 'SimplePie_Category');
